Pristine core 4.8 conf.

+++ seep-system/examples/acita_demo_2015/core-emane/ubuntu_trusty_setup.sh
@@ -44,27 +44,47 @@ popd
 
 #N.B. You need to install Oracle Java 1.7 manually first!
 #TODO: Automate java install
-
+if [ ! -f jdk-7u79-linux-x64.tar.gz ]; then
+    echo "JDK not found"
+    tar -xzvf jdk-7u79-linux-x64.tar.gz 
+    sudo cp -r jdk1.7.0_79 /usr/lib/jvm
+    sudo update-alternatives --install "/usr/bin/java" "java" "/usr/lib/jvm/jdk1.7.0_79/bin/java" 1
+    sudo update-alternatives --install "/usr/bin/javac" "javac" "/usr/lib/jvm/jdk1.7.0_79/bin/javac" 1
+    sudo update-alternatives --install "/usr/bin/jar" "jar" "/usr/lib/jvm/jdk1.7.0_79/bin/jar" 1
+    echo "Installed JDK, manual configuration of alternatives required, exiting."
+    exit 1
+fi
 
 echo "Downloading BonnMotion."
 #BonnMotion
 wget http://sys.cs.uos.de/bonnmotion/src/bonnmotion-2.1.3.zip
+unzip bonnmotion-2.1.3.zip
 
 # TODO: Automate bm install.
-
+#pushd bonnmotion-2.1.3
+#./install
+# TODO: Requires manual intervention to confirm java location
+#sudo ln -s `pwd`/bin/bm /usr/bin/bm
+#popd
 
 echo "Installing maven"
 #Maven (you'll perhaps need to manually install soot-framework-2.5.0.jar into the local mvn repository
 # libs/soot/soot-framework/2.5.0/soot-2.5.0.jar
 sudo apt-get install maven
-
-#TODO: using the suggested mvn install ... when mvn clean compile assembly:single fails).
+pushd ../../../../..
+mvn install:install-file -DgroupId=soot -DartifactId=soot-framework -Dversion=2.5.0 -Dpackaging=jar -Dfile=libs/soot/soot-framework/2.5.0/soot-2.5.0.jar
+./meander-bld.sh
+popd
 
 echo "Downloading NRL OLSR"
 #Nrlolsr (optional)
 wget http://downloads.pf.itd.nrl.navy.mil/olsr/nrlolsrdv7.8.1.tgz
-
-#TODO: Install nrlolsr
+tar -xzvf nrlolsrdv7.8.1.tgz
+sudo apt-get install libpcap-dev
+pushd nrlolsr/unix
+make -f Makefile.linux
+sudo ln -s `pwd`/nrlolsrd /usr/bin/nrlolsrd
+popd
 
 #olsrd 0.6.3 from ppa at https://launchpad.net/~guardianproject/+archive/ubuntu/commotion (Don't forget to update before installing!).
 echo "Downloading OLSRD"
@@ -73,6 +93,11 @@ wget "http://www.olsr.org/releases/0.9/$OLSRD_PKG"
 tar -xzvf $OLSRD_PKG  
 
 #TODO: Install OLSRD
+sudo apt-get install bison flex
+pushd olsrd-0.9.0.2
+make
+sudo make install
+popd
 
 echo "Installing pip + python packages."
 #For 14.04 can just install olsrd from apt-get directly.
